requirements.txt
pymongo
simplejson
xmltodict
python-dateutil
requests
pyyaml
cherrypy